+++ learnergy/models/rbm.py
@@ -454,7 +454,7 @@ class RBM(Model):
             # For every batch
             for samples, _ in batches:
                 # Flattening the samples' batch
-                samples = samples.view(len(samples), self.n_visible).float()
+                samples = samples.view(len(samples), self.n_visible)
 
                 # Checking whether GPU is avaliable and if it should be used
                 if self.device == 'cuda':
@@ -533,7 +533,7 @@ class RBM(Model):
         # For every batch
         for samples, _ in batches:
             # Flattening the samples' batch
-            samples = samples.view(len(samples), self.n_visible).float()
+            samples = samples.view(len(samples), self.n_visible)
 
             # Checking whether GPU is avaliable and if it should be used
             if self.device == 'cuda':
